Feat (account) : add user account management.

+++ src/Service/ImageService.php
@@ -6,7 +6,7 @@ class ImageService
 {
     private const MAX_FILE_SIZE = 5 * 1024 * 1024;
 
-    private const  ALLOWED_TYPES = [
+    private const ALLOWED_TYPES = [
             'image/jpeg' => 'jpg',
             'image/png' => 'png',
             'image/webp' => 'webp'
@@ -14,6 +14,9 @@ class ImageService
 
     private const UPLOAD_DIRECTORY =
         __DIR__ . '/../../public/assets/images/pictures-books/';
+
+    //Ajoute une image par defaut
+    private const DEFAULT_IMAGE = 'default-book.jpg';
 
     //Enregistre une image dans le dossier des livres
     public function upload(
@@ -24,7 +27,7 @@ class ImageService
             !isset($_FILES[$inputName])
             || $_FILES[$inputName]['error'] === UPLOAD_ERR_NO_FILE
         ) {
-            return null;
+            return self::DEFAULT_IMAGE;
         }
 
         if ($_FILES[$inputName]['error'] !== UPLOAD_ERR_OK) {
@@ -53,7 +56,6 @@ class ImageService
         $fileName = uniqid() . '.' . self::ALLOWED_TYPES[$fileType];
 
         $uploadPath = self::UPLOAD_DIRECTORY . $fileName;
-
         
         if (!move_uploaded_file(
             $_FILES[$inputName]['tmp_name'],
@@ -68,13 +70,19 @@ class ImageService
     //Supprime une image du dossier
     public function delete(?string $fileName): void 
     {
-        if (empty($fileName)) {
+        if (
+            empty($fileName)
+            || $fileName === self::DEFAULT_IMAGE
+        ) {
             return;
         }
 
         $filePath = self::UPLOAD_DIRECTORY . $fileName;
 
-        if (file_exists($filePath)) {
+        if (
+            file_exists($filePath)
+            && is_file($filePath)
+        ) {
             unlink($filePath);
         }
     }
